Add files via upload
The problem of insecure software is perhaps the most important technical challenge of our time. In this new paradigm, we find increasingly complex applications that support and manage huge amounts of business and process data from millions of users, sometimes as critical as personal and financial data. All this, together with more and more immediate development deadlines associated and new needs derived from new development methodologies, makes it necessary to simplify and jointly manage a framework of good practices and controls in order to build more and more resilient applications.
In recent decades, several attempts have been made to define the term vulnerability. In Software Security, we can summarize it as a possible weakness in our application that allows a malicious agent to take advantage of it. Obviously, the greater the value of the vulnerable asset (or assets), and the more frequent this threat occurs, the greater the risk to the business that our application represents. Therefore, during the initial phases of the SDLC (obtaining requirements and solution design within the SDLC), it is essential to have all these concepts present. Only from the beginning of security from design (Security by design), we can establish our application on the three pillars that are part of information security: Confidentiality, Integrity and Availability.
  
Source: NIST (https://www.nist.gov/ )

SSDLC (Secure Software Development Lifecycle) is this paradigm that seeks to redefine these processes and controls that allow to build safer applications by integrating them within the cycle, obtaining metrics in all phases of the development. Within this paradigm, a correct definition of security requirements and a correlation with the threats faced by our application constitute the base to apply this principle.